Provide a more approachable summary of Klavins pub
The content presented in Klavins et. al. (Graph Grammars for Self
Assembling Robotic Systems) is dense and difficult to digest. I am
providing a summary of the paper's definitions (with theorems and
examples forthcoming) to illustrate my own comprehension and to provide
context for other individuals who do not have a graduate-level
understanding of the mathematical underpinning of the paper.